requirements.txt
pandas
jinja2
requests
python-dotenv
